lightway-client: Register Ctrl-C handler later
... right before the main event loop begins.

I used the wrong IP address in a test and was therefore blocking before
reaching the mainloop, in `io::outside::Tcp::new`, I suppose TCP would
eventually have timed out but in the meantime I was unable to Ctrl-C out of it
since the handler had been registered but nothing was yet receiving from
`ctrlc_rx`. By defering the registration we leave the default OS behaviour
(term the process) in place until we are ready to deal with things ourselves.

It can be reproduced with the following if there is no server running:

```console
$ sudo ip netns exec lightway-client ./target/debug/lightway-client --config-file tests/client/client_config.yaml --mode tcp --ca-cert ca.crt --server 192.168.100.1:443
```

I don't think there is a similar failure mode for the server side, since it
receives rather than initiates connections, but make the same change there
nonetheless, there might be other similar issues with another step of the setup
process.