Switch from `derivative` to `educe`
The derivative crate appears to have become unmaintained, e.g.
https://github.com/mcarton/rust-derivative/issues/117.

`educe` appears to be the more popular looking of the two alternatives I found
there (the other was `derive-where`, 2.3M downloads vs 8.5M).

derivative has an apparent bug where `bounds=""` does not work with
`format_with`, I tripped over this while trying to add a field with a lifetime
parameter to `ClientConfig`, which resulted in a confusing `unused lifetime`
message for a lifetime which was used, just not for the field using
`format_with`.