tests: Add test cases with a single Tokio worker thread
The hope is that this will help catch potential deadlocks or other
problems/interactions which do not manifest if there are multiple worker
threads.

To facilitate this these tests enable PMTUD and keepalive so we have the large
set of tasks.

Also enable io_uring in one of the tests but not the other, in principal this
is independent of the tunnel mode so this allows us to test both io_uring and
tun modes without exploding the test matrix any further.