CVPN-939 Allow Custom Action for Event Callbacks
Currently event callbacks are contained in lightway-client's lib.rs with
no way of overriding, which make it impossible for components outside of
the client to react to event callbacks.

This change allows injection of a custom callback handler, a function
pointer, which will be called as part of the built-in handle_event
function.