lightway-core: Add `Event::TlsKeysUpdateCompleted` event
This means we must track when a key update is pending.

Note that `wolfssl::Session::is_update_keys_pending()` is calling
`wolfSSL_key_update_response` in the underlying library. This function is a
simple getter of an internal variable (IOW, cheap) so unlike in the C
lightway-core implementation we do not gate the call with our own condition.