lightway-core: Start key update timer only after online
... since we only rotate keys after we are online it is better not to include
the time we spent getting to the Online state in the initial interval.

Since the interval is on the order of minutes and the time to get online is on
the order of milliseconds the practical impact of this is neglibible. However
this will make a future change ("lightway-core: Encapsulate key update state
into a type.") closer to a "no semantic" change update.