lightway-{client,server}: Use a oneshot signal for the stop_signal (ctrl-c)
The receiver immediate exits when the channel is received, sending more
ctrl-c's will do nothing.

The `ctrlc::set_handler` API requires a `FnMut`, which is incompatible with
direct use of `oneshot::Sender` since `send` consumes, meaning the closure
would be `FnOnce`. We can use an `Option` and `take` since we only need to send
on the first Ctrl-C we get.