lightway-{client,server}: Failure to send Ctrl-C signal is not fatal
Failing to send on a oneshot channel can only happen if the other end has died,
which is already fatal (or soon will be).

Panicking in the Ctrl-C handler "will not be caught and will cause the signal
handler thread to stop" according to [`ctrlc::set_handler` docs][], so the
`.expect()` is not very useful.

[`ctrlc::set_handler` docs]: https://detegr.github.io/doc/ctrlc/fn.set_handler.html