CVPN-1144: add oneshot and continuous keepalives
In network change scenarios we want to be able to send
a few keepalives to ensure our connection is still working.

We adapt the existing keepalive infrastructure to suit this purpose
we want to be able to send a singleshot of keepalives but the
existing infra only allows for continuous.

We add a flag to specify either oneshot or continuous mode. Both
versions will continue to use the timeouts and intervals to specify
how often to send these keepalives but in the oneshot the keepalives
are disabled upon receiving a reply. Continuous keepalives however
remains the same.