CVPN-1144: add network change signals
We want to be able to signal to the client that a
network change is occuring. For this we add a signal that
can be used externally.

When these network change scenarios happen we want to disconnect
when connected via TCP. This is because TCP doesn't survive a
network change.

In UDP however we want to send keepalives to ensure that the
connection continues to function. UDP is expected to continue
to function after a network change.