client: make client apis public
So far, only lightway_client::client() api was public, which is a monolith.
This will start, connect and run all tasks within it.

By exposing more apis to public, we can enable external client application developers
to re-use modules which are relevant, and replace other modules and create their own clients.

This commit exposes some of these apis and structs