CVPN-942 Handle duplicate message as fatal error
The observed behaviour in client when this error happens is that the
connection will hang indefinitely until it is manually restarted. There
is nothing that the client can do to resurrect the failing state.

Handling the error as fatal will trigger early termination of the
client, which would allow a faster restart.