lightway-server: udp: Use `async_io` and `try_recv_buf_from`
In order to support `IP_PKTINFO` for incoming data (so we can identify the
local socket address where the data arrived) we need to switch to the `recvmsg`
API. Since this is not supported by the Tokio `UdpSocket` we will be switching
to using [`socket2`][] APIs over the raw fd.

Start by switching to the `async_io` interface. Since this just involves
switching to the `try_` variant of `recv_buf_from` within the callback there
should be no semantic change here.

[`socket2]: https://crates.io/crates/socket2