lightway-server: udp: Enable and use `IP_PKTINFO` to get local listening address
For now log but the intention is to pass this to the connection object where it
can be used e.g. for auth decisions for an instance which listens on `0.0.0.0`.

The `recvmsg()` API for control messages is pretty fiddly and involves a lot of
unsafe `libc` usage. This is encapsulated into a `cmsg` module to try and keep
the blast radius under control.

There is an alternative to `socket2` APIs in the [`nix`][] crate, which does
safely support the CMSG stuff, however that API uses `IoSliceMut` for the data
buffer which does not allow for passing an uninitialized buffer (`socket2` is
based around `MaybeUninitSlice`). Since this is a hot datapath piece of code we
prefer to avoid having to initialize read buffers to zero before filling them
and so take the unsafe hit ourselves.

Turn on various additional unsafe lints to try and keep to best practice.

[`nix]: https://crates.io/crates/nix