lightway-server: Pass only relevant information to auth callbacks
Create a view type over `ConnectionState` which contains the fields related to
auth and expose that as the expected type passed to the auth callback.

This avoids exposing the full details of `ConnectionState` to consumers of
lightway-server as a library crate since some of it is internal (and hard to
construct for mocking purposes in consumers).

The existing `AuthMetrics` type is repurposed as `AuthAdapter` to perform the
conversion.