lightway-core: Remove some unnecessary clones.
The ip pool one isn't terribly performance sensitive but the clone of
`inside_io` in `handle_outside_data_bytes` one is a (small) overhead on the
data path.

It seems that rustc has gotten better at spotting split borrows so the clone in
the ip pool case is no longer required.

In `handle_outside_data_bytes` the use of the helper function apparently
defeats this new ability, so switch to taking the reference directly.