lightway-server: udp: Crank a metric if received message is truncated
With the APIs used previously (`recv_from`) messages were silently truncated (a
property of the underlying syscall).

Now that we use `recvmsg` we can detect this (via the `MSG_TRUNC` bit in the
headers flags) and raise a metric. This should not be happening in practice
since we should be providing a large enough buffer. Continue to process the
packet, it will be discarded later during parse if it comes to that.

Ideally we would also like to check for `MSG_CTRUNC`, but this is not exposed
by the `socket2` abstraction. We will handle this possibility gracefully when
we come to parse the control message buffer.