lightway-server: ip_manager: Allocate a pool of free addresses up front
... rather than on demand as IPs are consumed.

This simplifies the code and will be easier to reason about as we work towards
supporting multiple IP pools.

As a consequence of this IP allocations are now somewhat random for a given
instance (as a property of the `HashSet` iterator).

This requires some test adjustments, in particular:
- tests can no longer rely on a particular allocation order
- the tests relating to "Consecutive Local and DNS IP" are no longer
  interesting because we are not expecting to allocate on demand in a fixed
  order.

Note that allocation is not completely random, it is still common to get back
the most recently inserted element on the next allocation. The
`ip_manager_inner_allocate_after_free_ip` test case was testing this behaviour,
but although it mostly seems to be the case it is not a given, in any case this
is not a terribly interesting property to be testing any more so this test is
removed.

The downside is an upfront allocation, for a /16 network this is approximately
255K.