lightway-server: ip_manager: Remove `allocated_ips` hash
We do not need both this and the free ips and this reduces worst case memory
use by half -- since neither hash can be relied on to shrink unless requested
they could both end up large enough to store all the IPs in the pool.

To prevent mistakes replace some of the checks on free with a check that the
freed IP belongs to the pool.