lightway-server: Add "ip_map" configuration option
When listening to 0.0.0.0 this allows mapping specific incoming IP addresses to
specific sub-ranges of the IP pool. SNAT in the backend can then be used to
choose an egress IP.

Note that the network and broadcast addresses of these sub-ranges are not
considered special/reserved, except for the first and last subrange which
contain the actual overall IP pool network and broadcast addresses. Other
reserved addresses are honoured throughout though.

The key of the map is an `IpAddr` rather than `Ipv4Addr`, even though we
currently wouldn't expect to see a `::V6` variant in practice. This is because
using the more general `IpAddr` allows for easier use with the `SocketAddr`
which is typically used as the lookup (since `SocketAddr::ip(...) -> IpAddr`).
However the value of the map is still `Ipv4Net` since we still currently only
support IPv4 internal networks.