lightway-server: udp: Track `peer_addr` as `socket2::SockAddr`
As well as the existing `std::net::SocketAddr`.

We would like to switch to `socket2::SockRef` operations in `send` (in order to
use `sendmsg`) which requires a `socket2::SockRef` argument.

Other parts of the API require instead a `std::new::SocketAddr`.

Although this is slightly redundant, since both types contain the same
information, this is a hot I/O path and the conversions while cheap do not
appear to be free:
https://docs.rs/socket2/0.5.7/src/socket2/sockaddr.rs.html#276-306
https://docs.rs/socket2/0.5.7/src/socket2/sockaddr.rs.html#208-238