lightway-server: udp: Use `try_io` and `socket2::SockRef::send_to`
In order to support sending with a chosen source address (to match the address
the peer believes it is talking to) we are working towards using
`SockRef::sendmsg`.

This commit starts by switching from Tokio's I/O routines to the lower level
socket2 versions, but using the same syscall under the hood. There should be no
semantic change from doing this.