lightway-server: udp: Pass a `in_pktinfo` control message to `sendmsg`
This selects the source IP address to use for the packet, which should be the
same address as the incoming request arrived on.

As with 25d38b3f45fb ("lightway-server: udp: Enable and use `IP_PKTINFO` to get
local listening address") this involves some unsafe code to initialize the
required control message data-structures. A new `cmsg::BufferBuilder`
abstraction is added to encapsulate this.

Running the enclosed tests under [`miri`][] raised some issued due to the
`CMSG_*` macros accessing `(*cmsg).cmsg_len as usize` with incorrect alignment.
It's unclear if this is a bug in the Rust libc or a requirement of the control
message APIs. Either way to avoid it we add an `align(16)` requirement to the
data type and do so for the `recvmsg` counterpart too.

Due to [`MaybeUninit::array_assume_init`][] currently being unstable we cannot
(easily? safely?) reuse the existing `Buffer` type, so add a new `BufferMut`
with an initialized array.

[`miri`]: https://github.com/rust-lang/miri
[`MaybeUninit::array_assume_init`]: https://doc.rust-lang.org/std/mem/union.MaybeUninit.html#method.array_assume_init