lightway-server: Do not instrument `UdpServer::data_received`
This is on the hot data path and instrumenting it is not terribly valuable, but
comes at a (small) performance cost.

Removing the instrumentation causes an iperf upload test to improve from ~1.85
Gbits/sec to ~1.98 Gbits/sec and a download test from ~2.21 Gbits/sec to ~2.24
Gbits/sec.

With the instrumentation gone clippy notices an unnecessary `return` so remove
that.