lightway-app-utils: iouring: Rework to push more processing into uring thread
This removes the need for the send and receive tasks posting sqe's to the
thread for it to post by doing that as part of the loop. It also pulls the cqe
processing into the same loop.

This now uses a model similar to that in most of the uring examples (in both C
and Rust).

This change improves an iperf upload test from 2.03 Gbits/sec to 2.19 Gbigs/sec
Gbits/sec and download from 2.34 Gbits/sec to 2.46 Gbits/sec.

The channel used for the tx and rx queues is changed from the async_channel to
tokio::mpsc. The performance difference appears to be in the noise, however we
no longer need the multi-consumer property (in fact we'd be happy with a spsc
channel if there was one) and we use tokio::mpsc everywhere else.