*: metrics: Register most metrics just once using `LazyLock`
Using the metrics macros appears to be doing non-trivial work on each use and
isn't just encapsulating a lazy initialization pattern.

The `tun_(to|from)_client` ones in particular are on the hot datapath and show
up in traces. Switching to the `LazyLock` scheme increases an iperf upload test
from 1.98 to 2.7 Gbits/sec and download 2.24 to 2.25 Gbits/sec (of which
`tun_*_client` are the major part).

We cannot do this for metrics which have various different labels (e.g. conn
created or link up) without adding a mapping of our own, since these are not
on the hotpath leave them using the macro each time.