tests: Ensure `net.ipv4.ip_forward` is set correctly in containers
Previously we relied on the system default, if this was 1 then everything was
ok, but if it was 0 then the middle container would not forward.

Configure everything explicitly:
* `lightway-remote`: 0
* `lightway-server`: 1
* `lightway-middle`: 1
* `lightway-client*`: 0

This renames the existing `setup_ip_forwarding` to `setup_ip_snat` to better
reflect what it does and to free up the `ip_forward` name to match the sysctl
name.