lightway-app-utils: Validate that config files are not world accessible
Since the file can contain credentials and other sensitive data this is
dangerous.

The new failure case is:
```console
Error: Invalid configuration file tests/server/server_config.yaml

Caused by:
    Configuration file is world accessible
```

This could have been made a warning, but the check is before we have configured
logging (since that may need the content of the config file, and may not result
in any logging), so take the easy path and make it a fatal error.

Sadly `git` does not track read and write permissions (only execute) so there
is no easy way to ensure that the test files have correctly limited permissions
for convenient dev testing. Add a note to the README and adjust the CI
configuration too.