lightway-app-utils: Use `fs_mistrust` to validate config file safety.
This covers way more edge cases than we could possibly hope to catch ourselves,
such as trusted files in untrusted directories.  See the [crate docs][] for all
the gory details.

In dev test this would complain because lightway (client and server) must be
run with sudo within the test containers but all the paths up to (and within)
the git checkout are owned by the developer not root. Allow this to be
addressed by providing the `LW_DANGEROUSLY_DISABLE_PERMISSIONS_CHECKS`
environment variable to disable the checks. This envar removes the need for
`chmod` for dev test, but it still remains in the e2e `Earthfile`s since they
run fully as root with root ownership so enforce the checks and otherwise fail
with e.g.

```
Error: Invalid configuration file server_config.yaml

Caused by:
    Incorrect permissions: /server_config.yaml is u=rw,g=rw,o=r; must be o-r
```

[crate docs]: https://docs.rs/fs-mistrust/latest/fs_mistrust/index.html