Ignore network unreachable socket error for linux-like OS when sending out packets from the udp socket of lightway-client.
If this error is not ignored, it will be passed to the WolfSSL layer. Then, the WolfSSL layer will reach a fatal socket error state and stop the inside IO loop during network change.
Ignoring this error to prevent the IO loop from exiting.
This change only applies to linux-like OS. Other OS like Windows may need a different implementation.