lightway-server: ip_pool: Avoid O(N) check on free
Avoid the O(N) call to `available_ips.contains` by keeping a `HashSet` of the
IP addresses we have handed out, which makes the test (which we combine with
`remove) take O(1) on average[^1] (worst case is still O(N)).

The `ip_manager::ip_pool::tests::lru_behaviour` test now takes an acceptable
amount of time (~0.25s).

With that we no longer need to bother with the check on ip pool since we know
we didn't ever hand out anything outside of that.

[^1]: https://doc.rust-lang.org/std/collections/index.html#maps