lightway-server: ip_pool: Allocate IPs to clients in LRU order
We currently store available IPs in a `HashSet`, this means that the current
effective (but unspecified) behaviour of the `HashSet` was to allocate the most
recently freed IP. This can make certain attacks based around guessing a
target's internal IP address somewhat easier to pull off.

Switch from using a `HashSet` to store to using a `VecDeque`, on allocate pop
from the front and on free push to the back which implements an LRU scheme.

This needs an O(N) call to `self.available_ips.contains` to preserve the
existing checks against freeing an IP outside of the set which the allocator
allocates. Because of this the new `ip_manager::ip_pool::tests::lru_behaviour`
test is slooooow (~20s).

For the tests we convert the `VecDeque` to a `HashSet` to facilitate comparison
against the expectations.