Relax permissions checks on configuration files
In 23c30e61c7e6 ("lightway-app-utils: Use `fs_mistrust` to validate config file
safety.") restrictions were added to ensure that configuration files and
secrets had permissions which would prevent another user from tampering with
them. This also required that the files were not readable by other users.

As of b8e2fc51e0ef ("lightway-server: Support multiple users") the
configuration file no longer contains any sensitive secrets so we can be more
tolerant of the file being readable.

We continue to enforce non-readableness for the server key and the user
database. We also add the missing checks for the server certificate, which we
allow to be readable.