lightway-server: reexport `ConnectionType` to applications
Since it is present in the public `ServerConfig` struct and we want consumers
to be able to set that without going via
`lightway_app_utils::args::ConnectionType` or taking a direct dependency on
`lightway_core` if they don't need to.

This was the only type used in `ServerConfig` which was not available from
either a reexport in the `lightway-server` crate or from `lightway-app-utils`.