app-utils: handle sq push error
There are cases where the sq.push() returns PushError instead of succeeding.
It is not a fatal error, and can be ignored safely.

When we encounter these errors in push_rx_events_to and push_tx_events_to, we add the
slot back to the available slots and restart to submit SQ.