app-utils: iouring: Use existing push_tx_events_to call in the loop flow
When we wait for tx_queue or cq_completion (event_fd), and receive a packet from
tx_queue, we tried to push that packet to SQ using push_one_tx_event_to. After that we
are trying to receive all packets and restart the loop.

But this is no different from the case, where the beginning of the loop we have some
packets in tx_queue. So we do not need different behavior for these two cases.

Just use the existing flow to add all packets from tx_queue to SQ