lightway-server: io: udp: Send reject replies from correct source address
If the server binds to `0.0.0.0` and has multiple possible egress IPs then we
must be explicit about the source address we want packets to come from.

For regular traffic we took care of this in 2cc7892f5260 ("lightway-server:
udp: Pass a `in_pktinfo` control message to `sendmsg`"), but the case where we
reject an unknown session was missed.

To achieve this we must switch to using `sendmsg` in the reject path, so that
we can provide the appropriate `IP_PKTINFO` control message.

Therefore refactor the body of `UdpSocket::send` into a helper function which
we can then use from `UdpServer::send_reject`. We could instead have
constructed a throw away `UdpSocket` but this would have involved cloning the
socket's `Arc` as well as constructing the appropriate `RwLock` for the
`peer_addr` field.