lightway-core: Ignore errors in `Connection::disconnect`
... when sending the goodbye message a failure could happen if the socket (for
a TCP connection) has already been closed by the peer (which is one reason we
might be disconnecting).

Similarly for the `try_shutdown` call which may try to send a TLS close
notification.

All callers already ignore the result of `Connection::disconnect`, so in
practice this is a noop change, but it more closely matches the "just a
courtesy" intention.