CVPN-1445 CVPN-1555 Test both ML-KEM and Kyber in e2e test
Validate that we do not break support clients which were built using Kyber in e2e test

Since we have added a new feature "kyber_only" in lightway-core that enables the client to use hybrid Kyber Level 5 instead of the hybrid ML-KEM 1024, we can test whether a server that supports both ML-KEM and Kyber can indeed support clients that are using either ML-KEM or Kyber.

We would need to build a new lightway client with "kyber_only" feature flag on. To do so, there's a new argument in earthly TEST that tells the client build container target to use/build the correct lightway client with "kyber_client" feature. The WolfSSL rust crate would then only build without any of the ML-KEM flags.