server: remove connections if not online by 60s
In some cases, the connection is becoming link-up, but never become online.
Since our inactive connection age out interval is 1 day, we are accumulating lot
of inactive connections and waste memory.

Introduce a new timeout STALE_AGE (60s) and disconnect the client,
if it is not online by that time.