core: build outside plugins only once for a connection
For outside plugins, we apply the plugins in two places.
1. In lightway_core::connection -> Ingress packets
2. In lightway_core::connection::WolfsslIoAdapter -> Egress packets

So we had to store the outside PluginList in both places.
But we were building PluginList twice from PluginFactory, instead of
using the same instance. This will case ingress and egress use different
instance of Plugin.

This commit fixes that by using the same instance in both places.