client: make udp socket writable on creation
With single threaded tokio runtime, the udp sockets return WouldBlock,
until we explicitly request the socket to be writable (UdpSocket::writable()).
The behavior seems different in multi threaded runtime, where it gets enabled
after the first few calls.

Since the outside io send is called from WolfSsl's callback which is not async,
we do not have a way to enable UdpSocket to be writable()

This causes the UDP lightway connection to fail, in case of single threaded environment.

This commit makes the socket to be writable as soon as it is created, so that
WolfSsl can use sync send directly successfully.

This behaviour is observed in client, but adding this fix to server since it seems logical.