server: Packet Codec
Packet Codec is supplied by lw-server to lw-core.

The packets in the encoder might have sit in the encoder for too long if
inside packets are not frequently received. To avoid having a high
packet delay,  lw-server triggers a flush when the encoder is ready.

There could be stale states in the decoder. The decoder's stale state is
cleaned periodically.

lw-server tracks the list of encoders and decoders for each connection.
Instead of accesing the codec via locking the connection, codec is access directly instead.
In this way, the number of locking to Connection is minimized.