core: move connection_created and connection_closed closer to Server::Connection
so it tracks the lifecycle of Server::Connection.

connection_created
- update it after we create the Connection successfully. Previously it was updated
  even if the Connection::new method fails.

connection_closed
- Update it only when the connection is getting dropped finally instead of calling from
Connection::disconnect(). Even when the connection is disconnection, there is no guarantee
that the connections are freed up since an Arc pointer could be stored somewhere else.
Updating it only on dropping the connection help to track those metrics consistently.