server: cleanup pending_session_id_rotations for disconnected clients
When we initate a session id rotation, we store the new session id along with
strong pointer of lightway_server::Connection inside the HashMap. We remove this
HashMap entry, when session rotation is completed.

In case the session rotation does not succeed, HashMap entry is not cleared
i.e. the strong pointer to Connection inside entry, is also not released
and thus Connection gets leaked.

There are two issues in the current approach:
1. Using strong pointer to Connection inside pending_session_id_rotations HashMap
2. Not cleaning up the pending_session_id_rotations HashMap periodically

First problem is fixed in previous commit by using Weak pointer to Connection in
pending_session_id_rotations.

This commit addresses the second issue by creating a task which periodically checks
for connection and cleansup if connection is already invalid/disconnected.