client: mock udp send success with non-fatal errors
For ConnectionRefused and NetworkUnreachable, we would like to wait for the
connection to come back up instead of fatally dropping the connection.

We need to mock that the send is success with the number of bytes consumed
same as the number of bytes requested to be sent.
Otherwise, WolfSSL perceives that no data is sent and try to send the same data again,
creating a live-lock until the connection resumes normally.