core: make inside io optional before connection
Inside io is not a mandatory parameter before starting the connection.
Making it optional will give us flexibility to test control plane without
testing traffic, eg. running multiple clients from same machine to test
load or control plan behavior.

One impact of this conditional checks for inside MTU during connection.
But this is wrong behavior and we should not rely on app to set the inside
mtu it should be from server. This will be fixed in further commit.