core: packet codec test in connection test
If codec is enabled in the test, the client sends a encoding request to the server before sending the test message to the server.
The logic is implemented by making the test client act as a state machine.

Note: codec is enabled in the test whenever Lightway UDP is used. The option to disable codec in Lightway UDP will be added in the later commits.