server: remove retrying logic for binding
We previously supported retrying logic if bind attempt fails. This is
to avoid any transient failures when a server instance is rotated or changed
and older instances are still alive for a brief moment.

This was a temporary workaround since we used different process manager
to manage C and Rust instances. Now that all the instances are running Rust
version, this workaround is not needed.

Also in the next commit, we add support to receive socket from external
application. This gives the external application a chance to bind and
use that socket in lightway server library.