client: remove inside_mtu as client argument
Tunnel mtu is a parameter which should be set from the server after
AuthSuccessWithConfigV4 message comes

Config message also contains tunnel ip addresses, but since we do
translation, it is fine to set static address to tunnel.
But for MTU, we cannot do that reliably and need to respect the
server set MTU. So move the mtu required check to runtime after
we get the network config message from server.