feat: let ServerAuth determine if inside packet encoding can be used
The default implementation allows inside packet encoding for all
connections, but the application can choose to override this behaviour
in its ServerAuth implementation

ServerAuthHandle is responsible for returning all available features for
the connection. In the future, a negotiation phase should be added to
allow Server to advertise which features are available to the Client.