client: allow stop signal when waiting for best connection
After #239, the client does not respond to stop signals until the first
server goes up. This causes the application to hang up and not respond
to SIGINT until the first connection connects or all connections fail.

For UDP connections especially, the connection does not fail
immediately, so the client can hang indefinitely until it receives a
SIGKILL.