app-utils: client: add cfg_aliases to simplify cfg
Now that we support multiple clients, conditional compilation
has become normal in client and app-utils.

It is posssible we miss in some places, if we keep on adding
cfg as long text.

Add cfg_aliases to build dependencies which simplifies cfg management.
We could extend this to include in sockopt cfg also. Will do it separate
PR.