client: add warning trace when connection task end
While adding the fatal error, I noticed the error is not
propagated to the main task, since we send only connected_signal
to find_best_connection.

This behavior is changed when we added parallel connect feature.
This gives no clue to the user about the error. So in best effort
try to print the error in warning trace.