keepalive: addd TracerDeltaExceeded message
Previously when the connection haven't heard from the server in a while,
we send a network change keepalive as a means to trigger a keepalive.
While it works, the naming is misleading, and we have to do extra
logging to differentiate the types of keepalives. The said logging
also introduced a lot of noise to the logs, which is the motivation of
this change.

Adding a new keepalive message to properly indicate the keepalive
intent.