app-utils: use add_address_v4 for tun address in windows
instead of set_network_address.

tun-rs crate in windows behave differently that it sets
the gateway if set_network_address is used.
This in turn makes Windows create default route with high priority.
And does not allow us to use our 0/1 and 128/1 route.

Ref:
- https://github.com/tun-rs/tun-rs/blob/7b38b0694283f9e6ed2ce6cc2f08da12ec10326e/src/platform/windows/device.rs#L296
- https://github.com/tun-rs/tun-rs/blob/7b38b0694283f9e6ed2ce6cc2f08da12ec10326e/src/platform/windows/netsh.rs#L99

Verified after this change, the default route created previously
was not there anymore.