client: Print correct best server index
Some server configurations are invalid, and fail to create a connection.
If this happens, the config is filtered out and ignored. When the best
connection is chosen, we print the index of the filtered array instead
of the original list.

This change fixes this by storing the original indices alongside the
connections, so we can print the original index.